+++ internal/models/eventAccounting.go
@@ -7,6 +7,7 @@ type EventAccounting struct {
 	EventAccountingEvent         primitive.ObjectID `bson:"event_accounting_event,omitempty" json:"event_accounting_event,omitempty"`
 	EventAccountingMessage       string             `bson:"event_accounting_message,omitempty" json:"event_accounting_message,omitempty"`
 	EventAccountingAmount        float64            `bson:"event_accounting_amount,omitempty" json:"event_accounting_amount,omitempty"`
+	EventAccountingPaidBy        primitive.ObjectID `bson:"event_accounting_paid_by,omitempty" json:"event_accounting_paid_by,omitempty"`
 	EventAccountingCreatedBy     primitive.ObjectID `bson:"event_accounting_created_by,omitempty" json:"event_accounting_created_by,omitempty"`
 	EventAccountingCreatedAt     primitive.DateTime `bson:"event_accounting_created_at,omitempty" json:"event_accounting_created_at,omitempty"`
 	EventAccountingUpdatedBy     primitive.ObjectID `bson:"event_accounting_updated_by,omitempty" json:"event_accounting_updated_by,omitempty"`